Force updating bad structures to better names
