+++ SpectatorTicketsExile3.0.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31729.503
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TutorialPlugin-EXILED", "TutorialPlugin-EXILED\TutorialPlugin-EXILED.csproj", "{4E7AF716-E3A4-4D2D-B0E6-C95EA31D0E8F}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SpectatorTickets", "SpectatorTicketsExile3.0\SpectatorTickets.csproj", "{4E7AF716-E3A4-4D2D-B0E6-C95EA31D0E8F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
