+++ SpectatorTicketsExile3.0/Commands/Example.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace TutorialPlugin_EXILED.Commands
+namespace SpectatorTickets_EXILED3.Commands
 {
     [CommandHandler(typeof(RemoteAdminCommandHandler))]
     class Allow : ICommand
@@ -18,10 +18,13 @@ namespace TutorialPlugin_EXILED.Commands
 
         public bool Execute(ArraySegment<string> arguments, ICommandSender sender, out string response)
         {
-
-            response = "Accepted command";
-            return true;
-
+            if(sender is Console)
+            {
+                response = "Updated plugin to do something if console";
+                return true;
+            }
+            response = "Updated plugin to do nothing if not console";
+            return false;
         }
     }
 }
