+++ SpectatorTicketsExile3.0/Handlers/Server.cs
@@ -7,14 +7,14 @@ using System.Text;
 using System.Threading.Tasks;
 
 
-namespace TutorialPlugin_EXILED.Handlers
+namespace SpectatorTickets_EXILED3.Handlers
 {
     class Server
     {
         public void OnWaitingForPlayers()
         {
-            Log.Info("Waiting for players.");
-            DebugLog.Log("Waiting for players???.");
+            //Log.Info("Waiting for players.");
+            //DebugLog.Log("Waiting for players???.");
         }
 
         public void OnRoundStarted()
