+++ SpectatorTicketsExile3.0/SpectatorTickets.csproj
@@ -7,8 +7,8 @@
     <ProjectGuid>{4E7AF716-E3A4-4D2D-B0E6-C95EA31D0E8F}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>TutorialPlugin_EXILED</RootNamespace>
-    <AssemblyName>TutorialPlugin-EXILED</AssemblyName>
+    <RootNamespace>SpectatorTickets_EXILE3</RootNamespace>
+    <AssemblyName>SpectatorTickets_EXILE3</AssemblyName>
     <TargetFrameworkVersion>v4.7.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <Deterministic>true</Deterministic>
@@ -102,14 +102,12 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Commands\Allow.cs" />
-    <Compile Include="Commands\Deny.cs" />
-    <Compile Include="Commands\HelloWorld.cs" />
+    <Compile Include="Commands\Example.cs" />
     <Compile Include="Config.cs" />
     <Compile Include="Handlers\Player.cs" />
     <Compile Include="Handlers\Server.cs" />
     <Compile Include="Handlers\SpectatorInfo.cs" />
-    <Compile Include="TutorialPlugin.cs" />
+    <Compile Include="Main.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup />
